proptypes and eslint

+++ part5/bloglist-frontend/src/components/Blogs.js
@@ -1,9 +1,9 @@
 import Blog from './Blog'
 import blogService from '../services/blogs'
-import React, { useState, useEffect, useRef} from 'react'
+import React, { useState, useEffect, useRef } from 'react'
 import Togglable from './Togglable'
 
-const LoggedStatus = ({user, setUser})=>{
+const LoggedStatus = ({ user, setUser }) => {
 
   const logOut = () =>
   {
@@ -17,7 +17,7 @@ const LoggedStatus = ({user, setUser})=>{
     <div>
       {user.name} logged in {' '}
       <button
-      onClick={()=>{logOut()}}
+        onClick = {() => {logOut()}}
       >logout</button>
       <br/>
       <br/>
@@ -25,12 +25,12 @@ const LoggedStatus = ({user, setUser})=>{
   );
 }
 
-const CreateBlog = ({user, setNotification, setError, blogs, setBlogs, hideForm})=>{
+const CreateBlog = ({ user, setNotification, setError, blogs, setBlogs, hideForm }) => {
   const [title, setTitle] = useState('');
   const [author, setAuthor] = useState('');
   const [url, setUrl] = useState('');
 
-  const  handleNewPost = async (event)=>{
+  const  handleNewPost = async (event) => {
     event.preventDefault();
     const newPost = {
       title: title,
@@ -40,22 +40,21 @@ const CreateBlog = ({user, setNotification, setError, blogs, setBlogs, hideForm}
     console.log(newPost);
     try{
       const response = await blogService.createNew(user, newPost)
-      console.log(response);
       setTitle('');
       setAuthor('');
       setUrl('');
-      setNotification("new blog created");
+      setNotification('new blog created');
       setError(false);
-      setTimeout(()=>{
+      setTimeout(() => {
         setNotification(null);
       }, 5000);
       setBlogs([...blogs, response]);
       hideForm();
     }
     catch (e){
-      setNotification("error creating blog");
+      setNotification('error creating blog');
       setError(true);
-      setTimeout(()=>{
+      setTimeout(() => {
         setNotification(null);
         setError(false);
       }, 5000);
@@ -65,28 +64,28 @@ const CreateBlog = ({user, setNotification, setError, blogs, setBlogs, hideForm}
   if (!user)
     return null;
   return (
-    <form onSubmit={(e)=>handleNewPost(e)}>
+    <form onSubmit = { (e) => handleNewPost(e) }>
       <div>
         title:
-        <input value={title} onChange={({target})=>setTitle(target.value)}></input>
+        <input value = {title} onChange = {({ target }) => setTitle(target.value)}></input>
       </div>
       <div>
         author:
-        <input value={author} onChange={({target})=>setAuthor(target.value)}></input>
+        <input value = {author} onChange = {({ target }) => setAuthor(target.value)}></input>
       </div>
       <div>
         url:
-        <input value={url} onChange={({target})=>setUrl(target.value)}></input>
+        <input value = {url} onChange = {({ target }) => setUrl(target.value)}></input>
       </div>
       <div>
-        <input type="submit" value="create"></input>
+        <input type = "submit" value = "create"></input>
       </div>
     </form>
 
   );
 }
 
-const Blogs = ({user, setUser, setNotification, setError})=>{
+const Blogs = ({ user, setUser, setNotification, setError }) => {
   const [blogs, setBlogs] = useState([])
   const formRef = useRef(null);
 
@@ -95,15 +94,58 @@ const Blogs = ({user, setUser, setNotification, setError})=>{
       return ;
     blogService.getAll().then(blogs =>
       setBlogs( blogs )
-    )  
+    )
   }, [user])
 
-  const deleteBlog = async(blog)=>{
+  useEffect(() => {
+    //sort blogs if not sorted
+    if (!blogs || blogs.length === 0)
+      return ;
+    let sorted = true;
+    blogs.forEach((b,i) => {
+      if (i >= blogs.length - 1)
+        return;
+      if (blogs[i+1].likes > b.likes)
+        sorted = false;
+    })
+    if (sorted)
+      return ;
+    const blogscpy = [...blogs];
+    blogscpy.sort((a,b) => b.likes - a.likes);
+    setBlogs(blogscpy);
+  },[blogs])
+
+  const likeBlog = async (blog) => {
+    try{
+      const newBlog = { ...blog };
+      newBlog.user = user.id;
+      newBlog.likes += 1;
+      await blogService.likePost(user, newBlog);
+      const blogsCpy = [...blogs];
+      blogsCpy.forEach((b) => {
+        if (b.id === blog.id)
+          b.likes += 1;
+      });
+      setBlogs(blogsCpy);
+    }
+    catch(e){
+      setNotification('error liking post');
+      setError(true);
+      setTimeout(() => {
+        setNotification(null);
+        setError(false);
+      }, 5000);
+    }
+  }
+
+  const deleteBlog = async(blog) => {
+    if (!window.confirm(`Delete ${blog.title}?`))
+      return ;
     try{
       const response = await blogService.deletePost(user,blog.id);
       console.log(response);
       let index = null;
-      blogs.forEach((b, i)=>{
+      blogs.forEach((b, i) => {
         if (b.id === blog.id)
           index = i;
       })
@@ -117,7 +159,7 @@ const Blogs = ({user, setUser, setNotification, setError})=>{
     catch(e){
       setNotification('Error deleting post');
       setError(true);
-      setTimeout(()=>{
+      setTimeout(() => {
         setError(false);
         setNotification(null);
       },5000);
@@ -130,23 +172,24 @@ const Blogs = ({user, setUser, setNotification, setError})=>{
   return (
     <>
       <h2>blogs</h2>
-      <LoggedStatus user={user} setUser={setUser}/>
-      <Togglable buttonLabel='create new' ref={formRef} cancelButton={true}>
+      <LoggedStatus user = {user} setUser = {setUser}/>
+      <Togglable buttonLabel = 'create new' ref = {formRef} cancelButton = {true}>
         <CreateBlog
-        user={user}
-        setNotification={setNotification}
-        setError={setError}
-        blogs={blogs}
-        setBlogs={(arg)=>setBlogs(arg)}
-        hideForm={()=>formRef.current.toggleHidden()}
-        />
+          user = {user}
+          setNotification = {setNotification}
+          setError = {setError}
+          blogs = {blogs}
+          setBlogs = {(arg) => setBlogs(arg)}
+          hideForm = {() => formRef.current.toggleHidden()}/>
       </Togglable>
       {blogs.map(blog =>
-        <Blog key={blog.id}
-        blog={blog}
-        deleteBlog={(blog)=>deleteBlog(blog)}/>
+        <Blog
+          key = {blog.id}
+          blog = {blog}
+          deleteBlog = {(blog) => deleteBlog(blog)}
+          likeBlog = {(blog) => likeBlog(blog)}/>
       )}
-     </>
+    </>
   );
 }
